+++ app/api/api_root.rb
@@ -94,6 +94,7 @@ class ApiRoot < Grape::API
   mount TutorialEnrolmentsApi
   mount UnitRolesApi
   mount UnitsApi
+  mount TutorNotesApi
 
   mount D2lIntegrationApi::D2lApi
   mount D2lIntegrationApi::OauthPublicApi
@@ -103,6 +104,7 @@ class ApiRoot < Grape::API
   mount WebcalPublicApi
   mount MarkingSessionsApi
   mount DiscussionPromptsApi
+  mount OverseerStepsApi
 
   mount Feedback::FeedbackChipApi
 
@@ -152,6 +154,8 @@ class ApiRoot < Grape::API
   AuthenticationHelpers.add_auth_to Feedback::FeedbackChipApi
   AuthenticationHelpers.add_auth_to MarkingSessionsApi
   AuthenticationHelpers.add_auth_to DiscussionPromptsApi
+  AuthenticationHelpers.add_auth_to OverseerStepsApi
+  AuthenticationHelpers.add_auth_to TutorNotesApi
 
   add_swagger_documentation \
     base_path: nil,