+++ app/api/entities/task_definition_entity.rb
@@ -19,6 +19,14 @@ module Entities
       expose :target_date
       expose :due_date
       expose :start_date
+      # expose :p_target_date, expose_nil: false
+      expose :c_target_date, expose_nil: false
+      expose :d_target_date, expose_nil: false
+      expose :hd_target_date, expose_nil: false
+
+      expose :c_start_date, expose_nil: false
+      expose :d_start_date, expose_nil: false
+      expose :hd_start_date, expose_nil: false
     end
 
     expose :upload_requirements, expose_nil: false do |task_definition, options|
@@ -50,7 +58,8 @@ module Entities
     expose :is_graded
     expose :max_quality_pts
     expose :overseer_image_id, if: ->(unit, options) { staff?(options[:my_role]) }, expose_nil: false
-    expose :assessment_enabled, if: ->(unit, options) { staff?(options[:my_role]) }
+    # expose :assessment_enabled, if: ->(unit, options) { staff?(options[:my_role]) }
+    expose :assessment_enabled
     expose :similarity_language, if: ->(unit, options) { staff?(options[:my_role]) }, expose_nil: false
     expose :assess_in_portfolio_only
     expose :use_resources_for_jplag_base_code, if: ->(unit, options) { staff?(options[:my_role]) }
@@ -61,5 +70,12 @@ module Entities
     expose :discussion_prompts_count do |task_def|
       task_def.discussion_prompts.size
     end
+
+    # expose :overseer_steps, using: OverseerStepEntity, if: ->(unit, options) { staff?(options[:my_role]) }
+    expose :overseer_steps, using: OverseerStepEntity do |task_def, options|
+      task_def.overseer_steps # options[:my_role] is still available inside the entity
+    end
+    expose :overseer_resource_files, if: ->(task_def, options) { staff?(options[:my_role]) }
+
   end
 end